Working code!
I have written the assembly that makes this all work. It may not be super
optimised just yet but by the time that I am done with it it will work at a
lightning pace. I will try and add more features to the code so that it can
display any results that it discovers. This should be a much valued feature
by the time it is done.

I have also added in files that should have been in the git repository from
the start.

I also need to write the solver code because currently all I have is code
that will run an assembly program that it is given.