Nearly perfectly working code.
This is pretty exciting. Soon I will have code that I can actually use to
accomplish this task. If that is complete then I will mail it to Richard
Buckland for the fun and joy of it. That will show that will probably make
everybody very happy and it would be awesome to watch. An actual practical
experiment.