Working runnable project.
I am nearly at the point where I can attempt to get the chip to solve any
length problem that I throw at it. Once that is complete the only thing left
to do will be to write some LCD code so that I can display my results to the
screen. That will probably be the most difficult part of this entire
endevour.

At any rate I can run a program and tell how long it was and wether or not
it was good to run it.